LF-3943 Add Storybook; clean up component; rename container
