+++ packages/webapp/src/stories/ReleaseBadge/styles.module.scss
@@ -1,0 +1,36 @@
+/*
+ *  Copyright 2023 LiteFarm.org
+ *  This file is part of LiteFarm.
+ *
+ *  LiteFarm is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  LiteFarm is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
+ */
+
+.wrapper {
+  position: relative;
+  transform: translateZ(0); // Create context for position: fixed
+  height: 300px;
+  outline: 1px solid black;
+  padding: 24px;
+}
+
+// Fixing issues only seen within Storybook UI
+.badge {
+  left: calc(50% - 200px);
+
+  p {
+    color: #426f57;
+  }
+
+  @media (max-width: 599px) {
+    left: 0px;
+    max-width: 100%;
+  }
+}
